Revert "Transaction Simulation Cleanup (#5977)"
This reverts commit 254e06db205c247d7472ba38278142de81c3d52e.
