+++ globals.d.ts
@@ -109,5 +109,4 @@ declare module 'react-native-dotenv' {
   export const REACT_NATIVE_RUDDERSTACK_WRITE_KEY: string;
   export const RUDDERSTACK_DATA_PLANE_URL: string;
   export const SILENCE_EMOJI_WARNINGS: boolean;
-  export const MWP_ENCRYPTION_KEY: string;
 }
