+++ package.json
@@ -70,7 +70,6 @@
     "@bradgarropy/use-countdown": "1.4.1",
     "@candlefinance/faster-image": "1.6.2",
     "@capsizecss/core": "3.0.0",
-    "@coinbase/mobile-wallet-protocol-host": "0.1.7",
     "@ensdomains/address-encoder": "0.2.16",
     "@ensdomains/content-hash": "2.5.7",
     "@ensdomains/eth-ens-namehash": "2.0.15",
