+++ src/App.tsx
@@ -1,11 +1,9 @@
 import './languages';
 import * as Sentry from '@sentry/react-native';
-import React, { lazy, Suspense, useCallback, useEffect, useState } from 'react';
-import { AppRegistry, Dimensions, LogBox, StyleSheet, View } from 'react-native';
-import { MobileWalletProtocolProvider } from '@coinbase/mobile-wallet-protocol-host';
-import { DeeplinkHandler } from '@/components/DeeplinkHandler';
-import { AppStateChangeHandler } from '@/components/AppStateChangeHandler';
-import { useApplicationSetup } from '@/hooks/useApplicationSetup';
+import React, { useCallback, useEffect, useRef, useState } from 'react';
+import { AppRegistry, AppState, AppStateStatus, Dimensions, InteractionManager, Linking, LogBox, View } from 'react-native';
+import branch from 'react-native-branch';
+
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import { enableScreens } from 'react-native-screens';
@@ -17,16 +15,26 @@ import { OfflineToast } from './components/toasts';
 import { designSystemPlaygroundEnabled, reactNativeDisableYellowBox, showNetworkRequests, showNetworkResponses } from './config/debug';
 import monitorNetwork from './debugging/network';
 import { Playground } from './design-system/playground/Playground';
+import handleDeeplink from './handlers/deeplinks';
+import { runWalletBackupStatusChecks } from './handlers/walletReadyEvents';
 import RainbowContextWrapper from './helpers/RainbowContext';
+import isTestFlight from './helpers/isTestFlight';
 import * as keychain from '@/model/keychain';
+import { loadAddress } from './model/wallet';
 import { Navigation } from './navigation';
 import RoutesComponent from './navigation/Routes';
+import { PerformanceContextMap } from './performance/PerformanceContextMap';
+import { PerformanceTracking } from './performance/tracking';
+import { PerformanceMetrics } from './performance/tracking/types/PerformanceMetrics';
 import { PersistQueryClientProvider, persistOptions, queryClient } from './react-query';
 import store from './redux/store';
+import { walletConnectLoadState } from './redux/walletconnect';
 import { MainThemeProvider } from './theme/ThemeContext';
+import { branchListener } from './utils/branch';
 import { addressKey } from './utils/keychainConstants';
 import { SharedValuesProvider } from '@/helpers/SharedValuesContext';
 import { InitialRoute, InitialRouteContext } from '@/navigation/initialRoute';
+import Routes from '@/navigation/routesNames';
 import { Portal } from '@/react-native-cool-modals/Portal';
 import { NotificationsHandler } from '@/notifications/NotificationsHandler';
 import { analyticsV2 } from '@/analytics';
@@ -34,16 +42,18 @@ import { getOrCreateDeviceId, securelyHashWalletAddress } from '@/analytics/util
 import { logger, RainbowError } from '@/logger';
 import * as ls from '@/storage';
 import { migrate } from '@/migrations';
+import { initListeners as initWalletConnectListeners } from '@/walletConnect';
+import { saveFCMToken } from '@/notifications/tokens';
 import { initializeReservoirClient } from '@/resources/reservoir/client';
 import { ReviewPromptAction } from '@/storage/schema';
+import { handleReviewPromptAction } from '@/utils/reviewAlert';
 import { initializeRemoteConfig } from '@/model/remoteConfig';
 import { NavigationContainerRef } from '@react-navigation/native';
 import { RootStackParamList } from './navigation/types';
 import { Address } from 'viem';
 import { IS_DEV } from './env';
+import { checkIdentifierOnLaunch } from './model/backup';
 import { prefetchDefaultFavorites } from './resources/favorites';
-
-const LazyRoutesComponent = lazy(() => import('./navigation/Routes'));
 
 if (IS_DEV) {
   reactNativeDisableYellowBox && LogBox.ignoreAllLogs();
@@ -52,40 +62,123 @@ if (IS_DEV) {
 
 enableScreens();
 
-const sx = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-});
+const containerStyle = { flex: 1 };
 
 interface AppProps {
   walletReady: boolean;
 }
 
 function App({ walletReady }: AppProps) {
-  const { initialRoute } = useApplicationSetup();
+  const [appState, setAppState] = useState(AppState.currentState);
+  const [initialRoute, setInitialRoute] = useState<InitialRoute>(null);
+  const eventSubscription = useRef<ReturnType<typeof AppState.addEventListener> | null>(null);
+  const branchListenerRef = useRef<ReturnType<typeof branch.subscribe> | null>(null);
+  const navigatorRef = useRef<NavigationContainerRef<RootStackParamList> | null>(null);
+
+  const setupDeeplinking = useCallback(async () => {
+    const initialUrl = await Linking.getInitialURL();
+
+    branchListenerRef.current = await branchListener(url => {
+      logger.debug(`[App]: Branch: listener called`, {}, logger.DebugContext.deeplinks);
+      try {
+        handleDeeplink(url, initialRoute);
+      } catch (error) {
+        if (error instanceof Error) {
+          logger.error(new RainbowError(`[App]: Error opening deeplink`), {
+            message: error.message,
+            url,
+          });
+        } else {
+          logger.error(new RainbowError(`[App]: Error opening deeplink`), {
+            message: 'Unknown error',
+            url,
+          });
+        }
+      }
+    });
+
+    if (initialUrl) {
+      logger.debug(`[App]: has initial URL, opening with Branch`, { initialUrl });
+      branch.openURL(initialUrl);
+    }
+  }, [initialRoute]);
+
+  const identifyFlow = useCallback(async () => {
+    const address = await loadAddress();
+    if (address) {
+      setTimeout(() => {
+        InteractionManager.runAfterInteractions(() => {
+          handleReviewPromptAction(ReviewPromptAction.TimesLaunchedSinceInstall);
+        });
+      }, 10_000);
+
+      InteractionManager.runAfterInteractions(checkIdentifierOnLaunch);
+    }
+
+    setInitialRoute(address ? Routes.SWIPE_LAYOUT : Routes.WELCOME_SCREEN);
+    PerformanceContextMap.set('initialRoute', address ? Routes.SWIPE_LAYOUT : Routes.WELCOME_SCREEN);
+  }, []);
+
+  const handleAppStateChange = useCallback(
+    (nextAppState: AppStateStatus) => {
+      if (appState === 'background' && nextAppState === 'active') {
+        store.dispatch(walletConnectLoadState());
+      }
+      setAppState(nextAppState);
+      analyticsV2.track(analyticsV2.event.appStateChange, {
+        category: 'app state',
+        label: nextAppState,
+      });
+    },
+    [appState]
+  );
 
   const handleNavigatorRef = useCallback((ref: NavigationContainerRef<RootStackParamList>) => {
+    navigatorRef.current = ref;
     Navigation.setTopLevelNavigator(ref);
   }, []);
 
+  useEffect(() => {
+    if (!__DEV__ && isTestFlight) {
+      logger.debug(`[App]: Test flight usage - ${isTestFlight}`);
+    }
+    identifyFlow();
+    eventSubscription.current = AppState.addEventListener('change', handleAppStateChange);
+
+    const p1 = analyticsV2.initializeRudderstack();
+    const p2 = setupDeeplinking();
+    const p3 = saveFCMToken();
+    Promise.all([p1, p2, p3]).then(() => {
+      initWalletConnectListeners();
+      PerformanceTracking.finishMeasuring(PerformanceMetrics.loadRootAppComponent);
+      analyticsV2.track(analyticsV2.event.applicationDidMount);
+    });
+
+    return () => {
+      eventSubscription.current?.remove();
+      branchListenerRef.current?.();
+    };
+  }, []);
+
+  useEffect(() => {
+    if (walletReady) {
+      logger.debug(`[App]: ✅ Wallet ready!`);
+      runWalletBackupStatusChecks();
+    }
+  }, [walletReady]);
+
   return (
     <Portal>
-      <View style={sx.container}>
+      <View style={containerStyle}>
         {initialRoute && (
           <InitialRouteContext.Provider value={initialRoute}>
-            <Suspense fallback={<View />}>
-              {/* @ts-expect-error - Property 'ref' does not exist on the 'IntrinsicAttributes' object */}
-              <LazyRoutesComponent ref={handleNavigatorRef} />
-            </Suspense>
+            <RoutesComponent ref={handleNavigatorRef} />
             <PortalConsumer />
           </InitialRouteContext.Provider>
         )}
         <OfflineToast />
       </View>
       <NotificationsHandler walletReady={walletReady} />
-      <DeeplinkHandler initialRoute={initialRoute} walletReady={walletReady} />
-      <AppStateChangeHandler walletReady={walletReady} />
     </Portal>
   );
 }
@@ -99,9 +192,9 @@ const AppWithRedux = connect<unknown, AppDispatch, unknown, RootState>(state => 
 }))(App);
 
 function Root() {
-  const [initializing, setInitializing] = useState(true);
+  const [initializing, setInitializing] = React.useState(true);
 
-  useEffect(() => {
+  React.useEffect(() => {
     async function initializeApplication() {
       await initializeRemoteConfig();
       await migrate();
@@ -207,21 +300,19 @@ function Root() {
             prefetchDefaultFavorites();
           }}
         >
-          <MobileWalletProtocolProvider secureStorage={ls.mwp} sessionExpiryDays={7}>
-            <SafeAreaProvider>
-              <MainThemeProvider>
-                <GestureHandlerRootView style={{ flex: 1 }}>
-                  <RainbowContextWrapper>
-                    <SharedValuesProvider>
-                      <ErrorBoundary>
-                        <AppWithRedux walletReady={false} />
-                      </ErrorBoundary>
-                    </SharedValuesProvider>
-                  </RainbowContextWrapper>
-                </GestureHandlerRootView>
-              </MainThemeProvider>
-            </SafeAreaProvider>
-          </MobileWalletProtocolProvider>
+          <SafeAreaProvider>
+            <MainThemeProvider>
+              <GestureHandlerRootView style={{ flex: 1 }}>
+                <RainbowContextWrapper>
+                  <SharedValuesProvider>
+                    <ErrorBoundary>
+                      <AppWithRedux walletReady={false} />
+                    </ErrorBoundary>
+                  </SharedValuesProvider>
+                </RainbowContextWrapper>
+              </GestureHandlerRootView>
+            </MainThemeProvider>
+          </SafeAreaProvider>
         </PersistQueryClientProvider>
       </RecoilRoot>
     </ReduxProvider>
