+++ src/handlers/deeplinks.ts
@@ -19,13 +19,6 @@ import { FiatProviderName } from '@/entities/f2c';
 import { getPoapAndOpenSheetWithQRHash, getPoapAndOpenSheetWithSecretWord } from '@/utils/poaps';
 import { queryClient } from '@/react-query';
 import { pointsReferralCodeQueryKey } from '@/resources/points';
-import { useMobileWalletProtocolHost } from '@coinbase/mobile-wallet-protocol-host';
-import { InitialRoute } from '@/navigation/initialRoute';
-
-interface DeeplinkHandlerProps extends Pick<ReturnType<typeof useMobileWalletProtocolHost>, 'handleRequestUrl' | 'sendFailureToClient'> {
-  url: string;
-  initialRoute: InitialRoute;
-}
 
 /*
  * You can test these deeplinks with the following command:
@@ -33,7 +26,7 @@ interface DeeplinkHandlerProps extends Pick<ReturnType<typeof useMobileWalletPro
  *    `xcrun simctl openurl booted "https://link.rainbow.me/0x123"`
  */
 
-export default async function handleDeeplink({ url, initialRoute, handleRequestUrl, sendFailureToClient }: DeeplinkHandlerProps) {
+export default async function handleDeeplink(url: string, initialRoute: any = null) {
   if (!url) {
     logger.warn(`[handleDeeplink]: No url provided`);
     return;
@@ -203,16 +196,6 @@ export default async function handleDeeplink({ url, initialRoute, handleRequestU
         logger.debug(`[handleDeeplink]: handling dapp`, { url });
         if (url) {
           Navigation.handleAction(Routes.DAPP_BROWSER_SCREEN, { url });
-        }
-        break;
-      }
-
-      case 'wsegue': {
-        const response = await handleRequestUrl(url);
-        if (response.error) {
-          // Return error to client app if session is expired or invalid
-          const { errorMessage, decodedRequest } = response.error;
-          await sendFailureToClient(errorMessage, decodedRequest);
         }
         break;
       }
