+++ src/navigation/Routes.ios.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable react/jsx-props-no-spreading */
-import { NavigationContainer, NavigationContainerRef } from '@react-navigation/native';
+import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 import React, { useContext } from 'react';
 import { AddCashSheet } from '../screens/AddCash';
@@ -104,7 +104,6 @@ import { useRemoteConfig } from '@/model/remoteConfig';
 import CheckIdentifierScreen from '@/screens/CheckIdentifierScreen';
 import { ControlPanel } from '@/components/DappBrowser/control-panel/ControlPanel';
 import { ClaimRewardsPanel } from '@/screens/points/claim-flow/ClaimRewardsPanel';
-import { RootStackParamList } from './types';
 
 const Stack = createStackNavigator();
 const NativeStack = createNativeStackNavigator();
@@ -293,13 +292,25 @@ function NativeStackNavigator() {
   );
 }
 
-const AppContainerWithAnalytics = React.forwardRef<NavigationContainerRef<RootStackParamList>, { onReady: () => void }>((props, ref) => (
-  <NavigationContainer onReady={props.onReady} onStateChange={onNavigationStateChange} ref={ref}>
-    <PointsProfileProvider>
-      <NativeStackNavigator />
-    </PointsProfileProvider>
-  </NavigationContainer>
-));
+const AppContainerWithAnalytics = React.forwardRef(
+  (
+    props: {
+      onReady: () => void;
+    },
+    ref
+  ) => (
+    <NavigationContainer
+      onReady={props.onReady}
+      onStateChange={onNavigationStateChange}
+      // @ts-ignore
+      ref={ref}
+    >
+      <PointsProfileProvider>
+        <NativeStackNavigator />
+      </PointsProfileProvider>
+    </NavigationContainer>
+  )
+);
 
 AppContainerWithAnalytics.displayName = 'AppContainerWithAnalytics';
