+++ src/navigation/config.tsx
@@ -29,7 +29,6 @@ import { BottomSheetNavigationOptions } from '@/navigation/bottom-sheet/types';
 import { Box } from '@/design-system';
 import { IS_ANDROID } from '@/env';
 import { SignTransactionSheetRouteProp } from '@/screens/SignTransactionSheet';
-import { RequestSource } from '@/utils/requestNavigationHandlers';
 
 export const sharedCoolModalTopOffset = safeAreaInsetValues.top;
 
@@ -278,7 +277,7 @@ export const signTransactionSheetConfig = {
   options: ({ route }: { route: SignTransactionSheetRouteProp }) => ({
     ...buildCoolModalConfig({
       ...route.params,
-      backgroundOpacity: route?.params?.source === RequestSource.WALLETCONNECT ? 1 : 0.7,
+      backgroundOpacity: route?.params?.source === 'walletconnect' ? 1 : 0.7,
       cornerRadius: 0,
       springDamping: 1,
       topOffset: 0,
