+++ src/redux/requests.ts
@@ -71,7 +71,7 @@ export interface WalletconnectRequestData extends RequestData {
 /**
  * Display details loaded for a request.
  */
-export interface RequestDisplayDetails {
+interface RequestDisplayDetails {
   /**
    * Data loaded for the request, depending on the type of request.
    */
@@ -154,7 +154,7 @@ export const addRequestToApprove =
           icons?: string[];
         }
   ) =>
-  async (dispatch: Dispatch<RequestsUpdateRequestsToApproveAction>, getState: AppGetState) => {
+  (dispatch: Dispatch<RequestsUpdateRequestsToApproveAction>, getState: AppGetState) => {
     const { requests } = getState().requests;
     const { walletConnectors } = getState().walletconnect;
     const { accountAddress, network, nativeCurrency } = getState().settings;
@@ -163,7 +163,7 @@ export const addRequestToApprove =
     const chainId = walletConnector._chainId;
     // @ts-expect-error "_accounts" is private.
     const address = walletConnector._accounts[0];
-    const displayDetails = await getRequestDisplayDetails(payload, nativeCurrency, chainId);
+    const displayDetails = getRequestDisplayDetails(payload, nativeCurrency, chainId);
     const oneHourAgoTs = Date.now() - EXPIRATION_THRESHOLD_IN_MS;
     // @ts-expect-error This fails to compile as `displayDetails` does not
     // always return an object with `timestampInMs`. Still, the error thrown
