+++ src/redux/walletconnect.ts
@@ -559,9 +559,7 @@ const listenOnNewMessages =
           return;
         }
         const { requests: pendingRequests } = getState().requests;
-        const request = !pendingRequests[requestId]
-          ? await dispatch(addRequestToApprove(clientId, peerId, requestId, payload, peerMeta))
-          : null;
+        const request = !pendingRequests[requestId] ? dispatch(addRequestToApprove(clientId, peerId, requestId, payload, peerMeta)) : null;
         if (request) {
           handleWalletConnectRequest(request);
           InteractionManager.runAfterInteractions(() => {
