+++ src/screens/WalletScreen/index.tsx
@@ -33,8 +33,6 @@ import { IS_ANDROID } from '@/env';
 import { RemoteCardsSync } from '@/state/sync/RemoteCardsSync';
 import { RemotePromoSheetSync } from '@/state/sync/RemotePromoSheetSync';
 import { UserAssetsSync } from '@/state/sync/UserAssetsSync';
-import { MobileWalletProtocolListener } from '@/components/MobileWalletProtocolListener';
-import { runWalletBackupStatusChecks } from '@/handlers/walletReadyEvents';
 
 const WalletPage = styled(Page)({
   ...position.sizeAsObject('100%'),
@@ -108,7 +106,6 @@ const WalletScreen: React.FC<any> = ({ navigation, route }) => {
     if (walletReady) {
       loadAccountLateData();
       loadGlobalLateData();
-      runWalletBackupStatusChecks();
     }
   }, [loadAccountLateData, loadGlobalLateData, walletReady]);
 
@@ -150,9 +147,6 @@ const WalletScreen: React.FC<any> = ({ navigation, route }) => {
         <UserAssetsSync />
         <RemoteCardsSync />
         <RemotePromoSheetSync />
-
-        {/* NOTE: This component listens for Mobile Wallet Protocol requests and handles them */}
-        <MobileWalletProtocolListener />
       </WalletPage>
     </View>
   );
