+++ src/storage/index.ts
@@ -3,7 +3,6 @@ import { MMKV } from 'react-native-mmkv';
 import { Account, Cards, Campaigns, Device, Review } from '@/storage/schema';
 import { EthereumAddress, RainbowTransaction } from '@/entities';
 import { Network } from '@/networks/types';
-import { SecureStorage } from '@coinbase/mobile-wallet-protocol-host';
 
 /**
  * Generic storage class. DO NOT use this directly. Instead, use the exported
@@ -13,8 +12,8 @@ export class Storage<Scopes extends unknown[], Schema> {
   protected sep = ':';
   protected store: MMKV;
 
-  constructor({ id, encryptionKey }: { id: string; encryptionKey?: string }) {
-    this.store = new MMKV({ id, encryptionKey });
+  constructor({ id }: { id: string }) {
+    this.store = new MMKV({ id });
   }
 
   /**
@@ -52,13 +51,6 @@ export class Storage<Scopes extends unknown[], Schema> {
   }
 
   /**
-   * Clear all values from storage
-   */
-  clear() {
-    this.store.clearAll();
-  }
-
-  /**
    * Remove many values from the same storage scope by keys
    *
    *   `removeMany([], [key])`
@@ -66,21 +58,6 @@ export class Storage<Scopes extends unknown[], Schema> {
    */
   removeMany<Key extends keyof Schema>(scopes: [...Scopes], keys: Key[]) {
     keys.forEach(key => this.remove([...scopes, key]));
-  }
-
-  /**
-   * Encrypt the storage with a new key
-   * @param newEncryptionKey - The new encryption key
-   */
-  encrypt(newEncryptionKey: string): void {
-    this.store.recrypt(newEncryptionKey);
-  }
-
-  /**
-   * Remove encryption from the storage
-   */
-  removeEncryption(): void {
-    this.store.recrypt(undefined);
   }
 }
 
@@ -111,27 +88,3 @@ export const cards = new Storage<[], Cards>({ id: 'cards' });
 export const identifier = new Storage<[], { identifier: string }>({
   id: 'identifier',
 });
-
-/**
- * Mobile Wallet Protocol storage
- *
- * @todo - fix any type here
- */
-const mwpStorage = new Storage<[], any>({ id: 'mwp', encryptionKey: process.env.MWP_ENCRYPTION_KEY });
-
-export const mwp: SecureStorage = {
-  get: async function <T>(key: string): Promise<T | undefined> {
-    const dataJson = mwpStorage.get([key]);
-    if (dataJson === undefined) {
-      return undefined;
-    }
-    return Promise.resolve(JSON.parse(dataJson) as T);
-  },
-  set: async function <T>(key: string, value: T): Promise<void> {
-    const encoded = JSON.stringify(value);
-    mwpStorage.set([key], encoded);
-  },
-  remove: async function (key: string): Promise<void> {
-    mwpStorage.remove([key]);
-  },
-};
