+++ src/utils/requestNavigationHandlers.ts
@@ -13,7 +13,7 @@ import {
 import { InteractionManager } from 'react-native';
 import { SEND_TRANSACTION } from './signingMethods';
 import { handleSessionRequestResponse } from '@/walletConnect';
-import ethereumUtils, { getNetworkFromChainId } from './ethereumUtils';
+import ethereumUtils from './ethereumUtils';
 import { getRequestDisplayDetails } from '@/parsers';
 import { RainbowNetworks } from '@/networks';
 import { maybeSignUri } from '@/handlers/imgix';
@@ -22,249 +22,8 @@ import { findWalletWithAccount } from '@/helpers/findWalletWithAccount';
 import { enableActionsOnReadOnlyWallet } from '@/config';
 import walletTypes from '@/helpers/walletTypes';
 import watchingAlert from './watchingAlert';
-import {
-  EthereumAction,
-  isEthereumAction,
-  isHandshakeAction,
-  PersonalSignAction,
-  RequestMessage,
-  useMobileWalletProtocolHost,
-} from '@coinbase/mobile-wallet-protocol-host';
-import { ChainId } from '@/__swaps__/types/chains';
-import { logger, RainbowError } from '@/logger';
-import { noop } from 'lodash';
-import { toUtf8String } from '@ethersproject/strings';
-import { BigNumber } from '@ethersproject/bignumber';
 
-export enum RequestSource {
-  WALLETCONNECT = 'walletconnect',
-  BROWSER = 'browser',
-  MOBILE_WALLET_PROTOCOL = 'mobile-wallet-protocol',
-}
-
-// Mobile Wallet Protocol
-
-interface HandleMobileWalletProtocolRequestProps
-  extends Omit<ReturnType<typeof useMobileWalletProtocolHost>, 'message' | 'handleRequestUrl' | 'sendFailureToClient'> {
-  request: RequestMessage;
-}
-
-const constructEthereumActionPayload = (action: EthereumAction) => {
-  if (action.method === 'eth_sendTransaction') {
-    const { weiValue, fromAddress, toAddress, actionSource, gasPriceInWei, ...rest } = action.params;
-    return [
-      {
-        ...rest,
-        from: fromAddress,
-        to: toAddress,
-        value: weiValue,
-      },
-    ];
-  }
-
-  return Object.values(action.params);
-};
-
-const supportedMobileWalletProtocolActions: string[] = [
-  'eth_requestAccounts',
-  'eth_sendTransaction',
-  'eth_signTypedData_v4',
-  'personal_sign',
-  'wallet_switchEthereumChain',
-];
-
-export const handleMobileWalletProtocolRequest = async ({
-  request,
-  fetchClientAppMetadata,
-  approveHandshake,
-  rejectHandshake,
-  approveAction,
-  rejectAction,
-  session,
-}: HandleMobileWalletProtocolRequestProps): Promise<boolean> => {
-  logger.debug(`Handling Mobile Wallet Protocol request: ${request.uuid}`);
-
-  const { selected } = store.getState().wallets;
-  const { accountAddress } = store.getState().settings;
-
-  const isReadOnlyWallet = selected?.type === walletTypes.readOnly;
-  if (isReadOnlyWallet && !enableActionsOnReadOnlyWallet) {
-    logger.debug('Rejecting request due to read-only wallet');
-    watchingAlert();
-    return Promise.reject(new Error('This wallet is read-only.'));
-  }
-
-  const handleAction = async (currentIndex: number): Promise<boolean> => {
-    const action = request.actions[currentIndex];
-    logger.debug(`Handling action: ${action.kind}`);
-
-    if (isHandshakeAction(action)) {
-      logger.debug(`Processing handshake action for ${action.appId}`);
-
-      const chainIds = RainbowNetworks.filter(network => network.enabled && network.networkType !== 'testnet').map(network => network.id);
-      const receivedTimestamp = Date.now();
-
-      const dappMetadata = await fetchClientAppMetadata();
-      return new Promise((resolve, reject) => {
-        const routeParams: WalletconnectApprovalSheetRouteParams = {
-          receivedTimestamp,
-          meta: {
-            chainIds,
-            dappName: dappMetadata?.appName || dappMetadata?.appUrl || action.appName || action.appIconUrl || action.appId || '',
-            dappUrl: dappMetadata?.appUrl || action.appId || '',
-            imageUrl: maybeSignUri(dappMetadata?.iconUrl || action.appIconUrl),
-            isWalletConnectV2: false,
-            peerId: '',
-            dappScheme: action.callback,
-            proposedChainId: request.account?.networkId ?? ChainId.mainnet,
-            proposedAddress: request.account?.address || accountAddress,
-          },
-          source: RequestSource.MOBILE_WALLET_PROTOCOL,
-          timedOut: false,
-          callback: async approved => {
-            if (approved) {
-              logger.debug(`Handshake approved for ${action.appId}`);
-              const success = await approveHandshake(dappMetadata);
-              resolve(success);
-            } else {
-              logger.debug(`Handshake rejected for ${action.appId}`);
-              await rejectHandshake('User rejected the handshake');
-              reject('User rejected the handshake');
-            }
-          },
-        };
-
-        Navigation.handleAction(
-          Routes.WALLET_CONNECT_APPROVAL_SHEET,
-          routeParams,
-          getActiveRoute()?.name === Routes.WALLET_CONNECT_APPROVAL_SHEET
-        );
-      });
-    } else if (isEthereumAction(action)) {
-      logger.debug(`Processing ethereum action: ${action.method}`);
-      if (!supportedMobileWalletProtocolActions.includes(action.method)) {
-        logger.error(new RainbowError(`[handleMobileWalletProtocolRequest]: Unsupported action type ${action.method}`));
-        await rejectAction(action, {
-          message: 'Unsupported action type',
-          code: 4001,
-        });
-        return false;
-      }
-
-      if (action.method === 'wallet_switchEthereumChain') {
-        const isSupportedChain = RainbowNetworks.find(network => network.id === BigNumber.from(action.params.chainId).toNumber());
-        if (!isSupportedChain) {
-          await rejectAction(action, {
-            message: 'Unsupported chain',
-            code: 4001,
-          });
-          return false;
-        }
-
-        await approveAction(action, { value: 'null' });
-        return true;
-      }
-
-      // NOTE: This is a workaround to approve the eth_requestAccounts action if the previous action was a handshake action.
-      const previousAction = request.actions[currentIndex - 1];
-      if (previousAction && isHandshakeAction(previousAction)) {
-        logger.debug('Approving eth_requestAccounts');
-        await approveAction(action, {
-          value: JSON.stringify({
-            chain: request.account?.chain ?? 'eth',
-            networkId: request.account?.networkId ?? ChainId.mainnet,
-            address: accountAddress,
-          }),
-        });
-        return true;
-      }
-
-      const nativeCurrency = store.getState().settings.nativeCurrency;
-
-      // @ts-expect-error - coinbase host protocol types are NOT correct e.g. {"data": [72, 101, 108, 108, 111, 32, 119, 111, 114, 108, 100], "type": "Buffer"}
-      if ((action as PersonalSignAction).params.message && (action as PersonalSignAction).params.message.type === 'Buffer') {
-        // @ts-expect-error - coinbase host protocol types are NOT correct e.g. {"data": [72, 101, 108, 108, 111, 32, 119, 111, 114, 108, 100], "type": "Buffer"}
-        const messageFromBuffer = toUtf8String(Buffer.from((action as PersonalSignAction).params.message.data, 'hex'));
-        (action as PersonalSignAction).params.message = messageFromBuffer;
-      }
-
-      const payload = {
-        method: action.method,
-        params: constructEthereumActionPayload(action),
-      };
-
-      const displayDetails = await getRequestDisplayDetails(payload, nativeCurrency, request.account?.networkId ?? ChainId.mainnet);
-
-      const requestWithDetails: RequestData = {
-        dappName: session?.dappName ?? session?.dappId ?? '',
-        dappUrl: session?.dappURL ?? '',
-        imageUrl: session?.dappImageURL ?? '',
-        address: (action as PersonalSignAction).params.address ?? accountAddress,
-        chainId: request.account?.networkId ?? ChainId.mainnet,
-        payload,
-        displayDetails,
-      };
-
-      return new Promise((resolve, reject) => {
-        const onSuccess = async (result: string) => {
-          logger.debug(`Ethereum action approved: [${action.method}]: ${result}`);
-          const success = await approveAction(action, { value: JSON.stringify(result) });
-          resolve(success);
-        };
-
-        const onCancel = async (error?: Error) => {
-          if (error) {
-            logger.debug(`Ethereum action rejected: [${action.method}]: ${error.message}`);
-            await rejectAction(action, {
-              message: error.message,
-              code: 4001,
-            });
-            reject(error.message);
-          } else {
-            logger.debug(`Ethereum action rejected: [${action.method}]: User rejected request`);
-            await rejectAction(action, {
-              message: 'User rejected request',
-              code: 4001,
-            });
-            reject('User rejected request');
-          }
-        };
-
-        Navigation.handleAction(Routes.CONFIRM_REQUEST, {
-          transactionDetails: requestWithDetails,
-          onSuccess,
-          onCancel,
-          onCloseScreen: noop,
-          chainId: request.account?.networkId ?? ChainId.mainnet,
-          address: accountAddress,
-          source: RequestSource.MOBILE_WALLET_PROTOCOL,
-        });
-      });
-    } else {
-      logger.error(new RainbowError(`[handleMobileWalletProtocolRequest]: Unsupported action type, ${action}`));
-      return false;
-    }
-  };
-
-  const handleActions = async (actions: typeof request.actions, currentIndex: number = 0): Promise<boolean> => {
-    if (currentIndex >= actions.length) {
-      logger.debug(`All actions completed successfully: ${actions.length}`);
-      return true;
-    }
-
-    logger.debug(`Processing action ${currentIndex + 1} of ${actions.length}`);
-    const success = await handleAction(currentIndex);
-    if (success) {
-      return handleActions(actions, currentIndex + 1);
-    } else {
-      // stop processing if an action fails
-      return false;
-    }
-  };
-
-  // start processing actions starting at index 0
-  return handleActions(request.actions);
-};
+export type RequestSource = 'walletconnect' | 'browser';
 
 // Dapp Browser
 
@@ -293,7 +52,7 @@ export const handleDappBrowserConnectionPrompt = (dappData: DappConnectionData):
         proposedChainId: dappData.chainId,
         proposedAddress: dappData.address,
       },
-      source: RequestSource.BROWSER,
+      source: 'browser',
       timedOut: false,
       callback: async (approved, approvedChainId, accountAddress) => {
         if (approved) {
@@ -318,31 +77,16 @@ export const handleDappBrowserConnectionPrompt = (dappData: DappConnectionData):
   });
 };
 
-const findWalletForAddress = async (address: string) => {
-  if (!address.trim()) {
-    return Promise.reject(new Error('Invalid address'));
-  }
-
+export const handleDappBrowserRequest = async (request: Omit<RequestData, 'displayDetails'>): Promise<string | Error> => {
   const { wallets } = store.getState().wallets;
-  const selectedWallet = findWalletWithAccount(wallets!, address);
-  if (!selectedWallet) {
-    return Promise.reject(new Error('Wallet not found'));
-  }
-
-  const isReadOnlyWallet = selectedWallet.type === walletTypes.readOnly;
+  const selectedWallet = findWalletWithAccount(wallets!, request.address);
+  const isReadOnlyWallet = selectedWallet!.type === walletTypes.readOnly;
   if (isReadOnlyWallet && !enableActionsOnReadOnlyWallet) {
     watchingAlert();
     return Promise.reject(new Error('This wallet is read-only.'));
   }
-
-  return selectedWallet;
-};
-
-export const handleDappBrowserRequest = async (request: Omit<RequestData, 'displayDetails'>): Promise<string | Error> => {
-  await findWalletForAddress(request.address);
-
   const nativeCurrency = store.getState().settings.nativeCurrency;
-  const displayDetails = await getRequestDisplayDetails(request.payload, nativeCurrency, request.chainId);
+  const displayDetails = getRequestDisplayDetails(request.payload, nativeCurrency, request.chainId);
 
   const requestWithDetails: RequestData = {
     ...request,
@@ -374,7 +118,7 @@ export const handleDappBrowserRequest = async (request: Omit<RequestData, 'displ
       onCloseScreen,
       chainId: request.chainId,
       address: request.address,
-      source: RequestSource.BROWSER,
+      source: 'browser',
     });
   });
 };
@@ -446,6 +190,6 @@ export const handleWalletConnectRequest = async (request: WalletconnectRequestDa
     network,
     address,
     chainId,
-    source: RequestSource.WALLETCONNECT,
+    source: 'walletconnect',
   });
 };
